requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
python-jose
passlib[bcrypt]
pydantic
python-multipart
requests